backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
pyairtable==3.2.0
python-dotenv==1.2.1
gunicorn==23.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
